Moving utils into shared directory
