+++ tests/unit/poller/controller/sources/ansible/data/inventory/invalid_inventory.json
@@ -3,7 +3,7 @@
         "hostvars": {
             "firewall01": {
                 "ansible_port": 22,
-                "ansible_ssh_private_key_file": "tests/unit/poller/controller/sources/ansible/data/inventory/keyfile",
+                "ansible_ssh_private_key_file": "tests/unit/poller/shared/sample_key",
                 "ansible_user": "vagrant",
                 "bgp_asn": 65533,
                 "default_vrf_asn": 65520,
