+++ tests/unit/poller/controller/sources/ansible/data/inventory/valid_inventory.json
@@ -4,7 +4,7 @@
             "firewall01": {
                 "ansible_host": "10.255.2.28",
                 "ansible_port": 22,
-                "ansible_ssh_private_key_file": "tests/unit/poller/controller/sources/ansible/data/inventory/keyfile",
+                "ansible_ssh_private_key_file": "tests/unit/poller/shared/sample_key",
                 "ansible_user": "vagrant",
                 "bgp_asn": 65533,
                 "default_vrf_asn": 65520,
@@ -98,7 +98,7 @@
                 "ansible_host": "10.255.2.184",
                 "ansible_network_os": "cumulus",
                 "ansible_port": 22,
-                "ansible_ssh_private_key_file": "tests/unit/poller/controller/sources/ansible/data/inventory/keyfile",
+                "ansible_ssh_private_key_file": "tests/unit/poller/shared/sample_key",
                 "ansible_user": "vagrant",
                 "anycast_gw_mac": "00:00:00:11:12:{{ item.vlanid }}",
                 "bgp_asn": "{{ ibgp_asn }}",
@@ -334,7 +334,7 @@
                 "ansible_host": "10.255.2.185",
                 "ansible_network_os": "cumulus",
                 "ansible_port": 22,
-                "ansible_ssh_private_key_file": "tests/unit/poller/controller/sources/ansible/data/inventory/keyfile",
+                "ansible_ssh_private_key_file": "tests/unit/poller/shared/sample_key",
                 "ansible_user": "vagrant",
                 "anycast_gw_mac": "00:00:00:11:12:{{ item.vlanid }}",
                 "bgp_asn": "{{ ibgp_asn }}",
