+++ tests/unit/poller/controller/sources/ansible/data/results/result.yaml
@@ -9,7 +9,7 @@ ansible-ns.10.255.2.184.22:
   port: 22
   transport: ssh
   username: vagrant
-  ssh_keyfile: tests/unit/poller/controller/sources/ansible/data/inventory/keyfile
+  ssh_keyfile: tests/unit/poller/shared/sample_key
   password: null
 ansible-ns.10.255.2.185.22:
   address: 10.255.2.185
@@ -22,7 +22,7 @@ ansible-ns.10.255.2.185.22:
   port: 22
   transport: ssh
   username: vagrant
-  ssh_keyfile: tests/unit/poller/controller/sources/ansible/data/inventory/keyfile
+  ssh_keyfile: tests/unit/poller/shared/sample_key
   password: null
 ansible-ns.10.255.2.28.22:
   address: 10.255.2.28
@@ -35,5 +35,5 @@ ansible-ns.10.255.2.28.22:
   port: 22
   transport: ssh
   username: vagrant
-  ssh_keyfile: tests/unit/poller/controller/sources/ansible/data/inventory/keyfile
+  ssh_keyfile: tests/unit/poller/shared/sample_key
   password: null
