+++ tests/unit/poller/controller/sources/native/data/inventory/valid_hosts.yaml
@@ -1,7 +1,7 @@
 - url: https://vagrant@192.168.123.252 devtype=eos
-- url: ssh://vagrant@192.168.123.232  keyfile=tests/unit/poller/controller/sources/native/data/inventory/keyfile
+- url: ssh://vagrant@192.168.123.232  keyfile=tests/unit/poller/shared/sample_key
 - url: https://vagrant@192.168.123.164 devtype=eos
 - url: ssh://192.168.123.70 username=admin password=admin
-- url: ssh://vagrant@192.168.123.230  keyfile=tests/unit/poller/controller/sources/native/data/inventory/keyfile
-- url: ssh://vagrant@192.168.123.54:2023  keyfile=tests/unit/poller/controller/sources/native/data/inventory/keyfile
+- url: ssh://vagrant@192.168.123.230  keyfile=tests/unit/poller/shared/sample_key
+- url: ssh://vagrant@192.168.123.54:2023  keyfile=tests/unit/poller/shared/sample_key
 - url: https://vagrant@192.168.123.123 password=vagrant devtype=panos
