+++ tests/unit/poller/controller/sources/native/data/results/results.yaml
@@ -34,7 +34,7 @@ native-ns.192.168.123.230.22:
   namespace: native-ns
   password: null
   port: 22
-  ssh_keyfile: tests/unit/poller/controller/sources/native/data/inventory/keyfile
+  ssh_keyfile: tests/unit/poller/shared/sample_key
   transport: ssh
   username: vagrant
 native-ns.192.168.123.232.22:
@@ -47,7 +47,7 @@ native-ns.192.168.123.232.22:
   namespace: native-ns
   password: null
   port: 22
-  ssh_keyfile: tests/unit/poller/controller/sources/native/data/inventory/keyfile
+  ssh_keyfile: tests/unit/poller/shared/sample_key
   transport: ssh
   username: vagrant
 native-ns.192.168.123.252.443:
@@ -73,7 +73,7 @@ native-ns.192.168.123.54.2023:
   namespace: native-ns
   password: null
   port: 2023
-  ssh_keyfile: tests/unit/poller/controller/sources/native/data/inventory/keyfile
+  ssh_keyfile: tests/unit/poller/shared/sample_key
   transport: ssh
   username: vagrant
 native-ns.192.168.123.70.22:
